+++ ft_printf/libft/source/custom/write_hex_upper.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_puthex_upper.c                                  :+:      :+:    :+:   */
+/*   write_hex_upper.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: donghyu2 <donghyu2@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/29 01:29:57 by donghyu2          #+#    #+#             */
-/*   Updated: 2022/11/29 01:54:25 by donghyu2         ###   ########.fr       */
+/*   Updated: 2022/12/02 01:44:33 by donghyu2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 static void	write_digit(short digit);
 
-void	write_hex_upper(unsigned long num)
+void	write_hex_upper(unsigned int num)
 {
 	if (num < 16)
 		write_digit(num & 0xf);
